python-opentelemetry-example/pyproject.toml
[tool.poetry]
name = "python-opentelemetry-example"
version = "0.1.0"
description = ""
authors = ["Uragami Taichi <backpaper0@gmail.com>"]
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = "^3.11"
fastapi = {extras = ["standard"], version = "^0.112.1"}
opentelemetry-instrumentation-fastapi = "^0.47b0"
opentelemetry-exporter-otlp = "^1.26.0"
python-dotenv = "^1.0.1"
langserve = {extras = ["server"], version = "^0.2.2"}
opentelemetry-instrumentation-langchain = "^0.27.0"
langchain-openai = "^0.1.22"
langchain = "^0.2.14"
aiohttp = "^3.10.4"
opentelemetry-instrumentation-aiohttp-client = "^0.47b0"


[tool.poetry.group.dev.dependencies]
taskipy = "^1.13.0"
flake8 = "^7.1.1"
black = "^24.8.0"
pylint = "^3.2.6"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.taskipy.tasks]
lint = "black app/ && pylint app/**/*.py && flake8 app/"